+++ documents/model.go
@@ -1,6 +1,7 @@
 package documents
 
 import (
+	"bytes"
 	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 	"github.com/centrifuge/go-centrifuge/errors"
 	"github.com/centrifuge/go-centrifuge/identity"
@@ -41,6 +42,11 @@ type CoreDocumentModel struct {
 	Document *coredocumentpb.CoreDocument
 }
 
+const (
+	// ErrZeroCollaborators error when no collaborators are passed
+	ErrZeroCollaborators = errors.Error("require at least one collaborator")
+)
+
 // NewCoreDocModel returns a new CoreDocumentModel
 // Note: collaborators and salts are to be filled by the caller
 func NewCoreDocModel() *CoreDocumentModel {
@@ -79,25 +85,28 @@ func (m *CoreDocumentModel) PrepareNewVersion(collaborators []string) (*CoreDocu
 	// copy read rules and roles
 	ncd.Roles = m.Document.Roles
 	ncd.ReadRules = m.Document.ReadRules
-	addCollaboratorsToReadSignRules(ncd, ucs)
-
+	err = ndm.addCollaboratorsToReadSignRules(ucs)
+	if err != nil {
+		return nil, err
+	}
 	err = ndm.fillSalts()
+
 	if err != nil {
 		return nil, err
 	}
 
 	if ocd.DocumentIdentifier == nil {
-		return nil, errors.New("coredocument.DocumentIdentifier is nil")
+		return nil, errors.New("Document.DocumentIdentifier is nil")
 	}
 	ncd.DocumentIdentifier = ocd.DocumentIdentifier
 
 	if ocd.CurrentVersion == nil {
-		return nil, errors.New("coredocument.CurrentVersion is nil")
+		return nil, errors.New("Document.CurrentVersion is nil")
 	}
 	ncd.PreviousVersion = ocd.CurrentVersion
 
 	if ocd.NextVersion == nil {
-		return nil, errors.New("coredocument.NextVersion is nil")
+		return nil, errors.New("Document.NextVersion is nil")
 	}
 
 	ncd.CurrentVersion = ocd.NextVersion
@@ -105,9 +114,19 @@ func (m *CoreDocumentModel) PrepareNewVersion(collaborators []string) (*CoreDocu
 	if ocd.DocumentRoot == nil {
 		return nil, errors.New("DocumentRoot is nil")
 	}
+	//how do we handle DocumentRoots
 	ncd.PreviousRoot = ocd.DocumentRoot
 
 	return ndm, nil
+}
+
+// AccountCanRead validate if the core document can be read by the account .
+// Returns an error if not.
+func (m *CoreDocumentModel) AccountCanRead(account identity.CentID) bool {
+	// loop though read rules
+	return m.findRole(coredocumentpb.Action_ACTION_READ_SIGN, func(role *coredocumentpb.Role) bool {
+		return isAccountInRole(role, account)
+	})
 }
 
 // FillSalts creates a new coredocument.Salts and fills it
@@ -116,10 +135,109 @@ func (m *CoreDocumentModel) fillSalts() error {
 	cd := m.Document
 	err := proofs.FillSalts(cd, salts)
 	if err != nil {
-		return errors.New("failed to fill coredocument salts: %v", err)
+		return errors.New("failed to fill Document salts: %v", err)
 	}
 
 	cd.CoredocumentSalts = salts
+	return nil
+}
+
+// initReadRules initiates the read rules for a given CoreDocumentModel.
+// Collaborators are given Read_Sign action.
+// if the rules are created already, this is a no-op.
+func (m *CoreDocumentModel) initReadRules (collabs []identity.CentID) error {
+	cd := m.Document
+	if len(cd.Roles) > 0 && len(cd.ReadRules) > 0 {
+		return nil
+	}
+
+	if len(collabs) < 1 {
+		return ErrZeroCollaborators
+	}
+
+	return m.addCollaboratorsToReadSignRules(collabs)
+}
+
+// addNewRule creates a new rule as per the role and action.
+func (m *CoreDocumentModel) addNewRule(role *coredocumentpb.Role, action coredocumentpb.Action) {
+	cd := m.Document
+	cd.Roles = append(cd.Roles, role)
+
+	rule := new(coredocumentpb.ReadRule)
+	rule.Roles = append(rule.Roles, role.RoleKey)
+	rule.Action = action
+	cd.ReadRules = append(cd.ReadRules, rule)
+}
+
+
+// findRole calls OnRole for every role,
+// if onRole returns true, returns true
+// else returns false
+func (m *CoreDocumentModel) findRole( action coredocumentpb.Action, onRole func(role *coredocumentpb.Role) bool) bool {
+	cd := m.Document
+	for _, rule := range cd.ReadRules {
+		if rule.Action != action {
+			continue
+		}
+
+		for _, rk := range rule.Roles {
+			role, err := getRole(rk, cd.Roles)
+			if err != nil {
+				// seems like roles and rules are not in sync
+				// skip to next one
+				continue
+			}
+
+			if onRole(role) {
+				return true
+			}
+
+		}
+	}
+
+	return false
+}
+
+// isAccountInRole returns true if account is in the given role as collaborators.
+func isAccountInRole(role *coredocumentpb.Role, account identity.CentID) bool {
+	for _, id := range role.Collaborators {
+		if bytes.Equal(id, account[:]) {
+			return true
+		}
+	}
+
+	return false
+}
+
+func getRole(key []byte, roles []*coredocumentpb.Role) (*coredocumentpb.Role, error) {
+	for _, role := range roles {
+		if utils.IsSameByteSlice(role.RoleKey, key) {
+			return role, nil
+		}
+	}
+
+	return nil, errors.New("role %d not found", key)
+}
+
+func (m *CoreDocumentModel) addCollaboratorsToReadSignRules (collabs []identity.CentID) error {
+	if len(collabs) == 0 {
+		return nil
+	}
+	// create a role for given collaborators
+	role := new(coredocumentpb.Role)
+	cd := m.Document
+	rk, err := utils.ConvertIntToByte32(len(cd.Roles))
+	if err != nil {
+		return err
+	}
+	role.RoleKey = rk[:]
+	for _, c := range collabs {
+		c := c
+		role.Collaborators = append(role.Collaborators, c[:])
+	}
+
+	m.addNewRule(role, coredocumentpb.Action_ACTION_READ_SIGN)
+
 	return nil
 }
